move all upgrade logic from rc1, rc2, and rc5 into v14.0.0 handler
