+++ app/upgrades/v14/migrations.go
@@ -1,7 +1,7 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
 
-package v14rc2
+package v14
 
 import (
 	"fmt"
