+++ app/upgrades/v14/migrations_test.go
@@ -1,10 +1,10 @@
-package v14rc2_test
+package v14_test
 
 import (
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/evmos/evmos/v14/app/upgrades/v14rc2"
+	"github.com/evmos/evmos/v14/app/upgrades/v14"
 	"github.com/evmos/evmos/v14/crypto/ethsecp256k1"
 	"github.com/evmos/evmos/v14/testutil"
 	testutiltx "github.com/evmos/evmos/v14/testutil/tx"
@@ -33,7 +33,7 @@ func (s *UpgradesTestSuite) TestUpdateMigrateNativeMultisigs() {
 
 	var (
 		migratedBalances sdk.Coins
-		migrationTarget  = v14rc2.NewTeamPremintWalletAcc
+		migrationTarget  = v14.NewTeamPremintWalletAcc
 		oldMultisigs     = make([]string, 0, len(affectedAccounts))
 	)
 
@@ -59,7 +59,7 @@ func (s *UpgradesTestSuite) TestUpdateMigrateNativeMultisigs() {
 	// Check validator shares before migration
 	expectedSharesMap := s.getDelegationSharesMap()
 
-	err := v14rc2.MigrateNativeMultisigs(s.ctx, s.app.BankKeeper, s.app.StakingKeeper, migrationTarget, oldMultisigs...)
+	err := v14.MigrateNativeMultisigs(s.ctx, s.app.BankKeeper, s.app.StakingKeeper, migrationTarget, oldMultisigs...)
 	s.Require().NoError(err, "failed to migrate native multisigs")
 
 	// Check that the multisigs have been updated
@@ -91,7 +91,7 @@ func (s *UpgradesTestSuite) TestInstantUnbonding() {
 	delegation, found := s.app.StakingKeeper.GetDelegation(s.ctx, s.address.Bytes(), s.validators[0].GetOperator())
 	s.Require().True(found, "delegation not found")
 
-	unbondAmount, err := v14rc2.InstantUnbonding(s.ctx, s.app.BankKeeper, s.app.StakingKeeper, delegation, s.bondDenom)
+	unbondAmount, err := v14.InstantUnbonding(s.ctx, s.app.BankKeeper, s.app.StakingKeeper, delegation, s.bondDenom)
 	s.Require().NoError(err, "failed to unbond")
 	s.Require().Equal(unbondAmount, math.NewInt(1e18), "expected different unbond amount")
