+++ app/upgrades/v14/setup_test.go
@@ -1,6 +1,6 @@
 // Copyright Tharsis Labs Ltd.(Evmos)
 // SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
-package v14rc2_test
+package v14_test
 
 import (
 	"testing"
