+++ app/upgrades/v14/vesting_test.go
@@ -1,8 +1,8 @@
-package v14rc2_test
+package v14_test
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/evmos/evmos/v14/app/upgrades/v14rc2"
+	"github.com/evmos/evmos/v14/app/upgrades/v14"
 	"github.com/evmos/evmos/v14/testutil"
 	"github.com/evmos/evmos/v14/x/vesting/types"
 )
@@ -11,18 +11,18 @@ func (s *UpgradesTestSuite) TestUpdateVestingFunders() {
 	s.SetupTest()
 
 	// Fund the affected accounts to initialize them and then create vesting accounts
-	s.prepareVestingAccount(v14rc2.VestingAddrByFunder1, v14rc2.OldFunder1)
-	for _, address := range v14rc2.VestingAddrsByFunder2 {
-		s.prepareVestingAccount(address, v14rc2.OldFunder2)
+	s.prepareVestingAccount(v14.VestingAddrByFunder1, v14.OldFunder1)
+	for _, address := range v14.VestingAddrsByFunder2 {
+		s.prepareVestingAccount(address, v14.OldFunder2)
 	}
 
 	// Run the upgrade function
-	err := v14rc2.UpdateVestingFunders(s.ctx, s.app.VestingKeeper, v14rc2.NewTeamPremintWalletAcc)
+	err := v14.UpdateVestingFunders(s.ctx, s.app.VestingKeeper, v14.NewTeamPremintWalletAcc)
 	s.Require().NoError(err, "failed to update vesting funders")
 
 	// Check that the vesting accounts have been updated
-	affectedAddrs := v14rc2.VestingAddrsByFunder2
-	affectedAddrs = append(affectedAddrs, v14rc2.VestingAddrByFunder1)
+	affectedAddrs := v14.VestingAddrsByFunder2
+	affectedAddrs = append(affectedAddrs, v14.VestingAddrByFunder1)
 	for _, address := range affectedAddrs {
 		accAddr := sdk.MustAccAddressFromBech32(address)
 		acc := s.app.AccountKeeper.GetAccount(s.ctx, accAddr)
@@ -32,7 +32,7 @@ func (s *UpgradesTestSuite) TestUpdateVestingFunders() {
 		s.Require().Equal(address, vestingAcc.Address, "expected different address in vesting account for %s", address)
 
 		// Check that the funder has been updated
-		s.Require().Equal(v14rc2.NewTeamPremintWalletAcc.String(), vestingAcc.FunderAddress, "expected different funder address for %s", address)
+		s.Require().Equal(v14.NewTeamPremintWalletAcc.String(), vestingAcc.FunderAddress, "expected different funder address for %s", address)
 	}
 }
