--- app/upgrades/v14rc5/constants.go
@@ -1,11 +1,0 @@
-// Copyright Tharsis Labs Ltd.(Evmos)
-// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/evmos/blob/main/LICENSE)
-
-package v14rc5
-
-const (
-	// UpgradeName is the shared upgrade plan name for mainnet
-	UpgradeName = "v14.0.0-rc5"
-	// UpgradeInfo defines the binaries that will be used for the upgrade
-	UpgradeInfo = `'{"binaries":{"darwin/arm64":"https://github.com/evmos/evmos/releases/download/v14.0.0-rc5/evmos_14.0.0-rc5_Darwin_arm64.tar.gz","darwin/amd64":"https://github.com/evmos/evmos/releases/download/v14.0.0-rc5/evmos_14.0.0-rc5_Darwin_amd64.tar.gz","linux/arm64":"https://github.com/evmos/evmos/releases/download/v14.0.0-rc5/evmos_14.0.0-rc5_Linux_arm64.tar.gz","linux/amd64":"https://github.com/evmos/evmos/releases/download/v14.0.0-rc5/evmos_14.0.0-rc5_Linux_amd64.tar.gz","windows/x86_64":"https://github.com/evmos/evmos/releases/download/v14.0.0-rc5/evmos_14.0.0-rc5_Windows_x86_64.zip"}}'`
-)
